backend/requirements.txt
fastapi~=0.97.0
injector~=0.20.1
starlette~=0.27.0
uvicorn~=0.22.0
aiopg~=1.4.0
retrying~=1.3.4
python-dotenv~=1.0.0
psycopg-binary==3.1.9
psycopg==3.1.9
pydantic~=1.10.9
humps~=0.2.2
aiofiles~=23.1.0
aiocache~=0.12.1
asyncpg~=0.27.0